// frontend/src/api/model/appConfigDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */
import type { AppConfigDtoType } from './appConfigDtoType';

export interface AppConfigDto {
  key: string;
  value: string;
  type: AppConfigDtoType;
  /** @nullable */
  category: string | null;
  /** @nullable */
  label: string | null;
  isPublic: boolean;
  updatedAt: string;
}